requirements.txt
beautifulsoup4==4.12.3
langchain==0.3.2
langchain_community==0.3.1
langchain_core==0.3.9
langchain_google_community==2.0.0
langchain_openai==0.2.2
Pillow==10.4.0
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.38.0
tqdm==4.66.5
pypdf
chromadb==0.4.14
rank_bm25
openpyxl
xlsxwriter
langchain-chroma
# pysqlite3-binary